장바구니, 위시리스트 수정

+++ Joya/src/com/joya/cart/controller/WishListToCartActionController.java
@@ -1,6 +1,8 @@
 package com.joya.cart.controller;
 
+import java.io.IOException;
 import java.io.UnsupportedEncodingException;
+import java.net.URLDecoder;
 import java.util.Map;
 
 import javax.servlet.ServletException;
@@ -19,59 +21,70 @@ import com.joya.wishlist.service.WishlistServiceImpl;
 
 
 public class WishListToCartActionController implements Controller {
-	
+
 	private WishlistService wishlistService = new WishlistServiceImpl();
-	
-	
+
+
 	@Override
 	public ModelAndView handleRequest(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, UnsupportedEncodingException {
 		ModelAndView mav = new ModelAndView();
-		
-		String[] productIds = request.getParameterValues("");
-		String email=null;
+
+		String[] productIds = request.getParameterValues("ji-chk");
+		String email=(String)request.getAttribute("email");
 		Map<String, String> cartMap=null;
 
 		Cookie[] cookies=request.getCookies();
-		for(Cookie cookie:cookies) {
-			if(cookie.getName().equalsIgnoreCase("cart")) {
-				cartMap=getCartMap(cookie.getValue());
-				break;
+		String cartCookieVal="";
+		if(cookies!=null){
+			System.out.println("쿠키의 갯수"+cookies.length);
+
+			for(Cookie cookie:cookies) {
+				System.out.println("쿠키이름 "+cookie.getName());
+				if(cookie.getName().equalsIgnoreCase("cart")) {
+					System.out.println("카트 쿠키 있음?");
+					cartCookieVal=URLDecoder.decode(cookie.getValue(), "utf-8");
+					break;
+				}
 			}
 		}
-		
-		for (String productId : productIds) {
-			cartMap.put(productId, "1");
-			wishlistService.delete(email, Integer.parseInt(productId));
+
+		cartMap=getCartMap(cartCookieVal);
+
+
+		if(productIds!=null) {
+			for (String productId : productIds) {
+				cartMap.put(productId, "1");
+				wishlistService.delete(email, Integer.parseInt(productId));
+			}
 		}
-		
-		
 		String cartInfo=mapToStrCart(cartMap);
 		cookies=request.getCookies();
 		for(Cookie cookie:cookies) {
 			if(cookie.getName().equalsIgnoreCase("cart")) {
+				System.out.println("카트 쿠키 있고 값 넣을거 :"+cartInfo);
 				cookie.setValue(cartInfo);
 				response.addCookie(cookie);
 				break;
 			}
 		}
-		
-		
-		mav.setView("/product/list.jsp");
-		
+
+		mav.setView("redirect:/cart.joya");
+
 		return mav;
 	}
-	
+
 	private Map<String, String> getCartMap(String cartCookieVal) {
 		Map<String, String> cartMap=new HashedMap<>();
 		String[] cartItems=cartCookieVal.split(Delimiter.CART_ITEM);
-		
-		for (String item : cartItems) {
-			String productId=item.split(Delimiter.CART_ITEM_INFO)[0];
-			String amount=item.split(Delimiter.CART_ITEM_INFO)[1];
-			cartMap.put(productId, amount);
+
+		if(!cartCookieVal.equals(" ")) {
+			for (String item : cartItems) {
+				String productId=item.split(Delimiter.CART_ITEM_INFO)[0];
+				String amount=item.split(Delimiter.CART_ITEM_INFO)[1];
+				cartMap.put(productId, amount);
+			}
 		}
-		
 		return cartMap;
 	}
